moved both argument types to one `nbt.rs` file
