+++ pumpkin/src/command/argument_types/mod.rs
@@ -224,8 +224,7 @@ pub mod entity_anchor;
 pub mod entity_selector;
 pub mod game_profile;
 pub mod identifier;
-pub mod nbt_compound_tag;
-pub mod nbt_tag;
+pub mod nbt;
 pub mod range;
 pub mod time;
 pub mod uuid;
